# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Containr API
  description: |
    Self-hosted PaaS on Proxmox - Railway-like platform for container orchestration and deployment.
    
    ## Overview
    Containr provides a comprehensive platform for deploying, managing, and scaling containerized applications on Proxmox infrastructure. This API allows you to manage projects, services, deployments, databases, and more through a RESTful interface.
    
    ## Authentication
    The API uses JWT-based authentication. Include your JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Base URL
    ```
    https://your-containr-domain.com/api/v1
    ```
  version: 1.0.0
  contact:
    name: Containr Documentation
    url: https://github.com/your-org/containr
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://your-containr-domain.com/api/v1
    description: Production server
paths:
...
  /services/{id}:
    get:
      tags:
        - Services
      summary: Get service
      description: Retrieve a specific service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      responses:
        '200':
          description: Service retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'

    put:
      tags:
        - Services
      summary: Update service
      description: Update a specific service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        '200':
          description: Service updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'

    delete:
      tags:
        - Services
      summary: Delete service
      description: Delete a specific service
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Service ID
          schema:
            type: string
      responses:
        '204':
          description: Service deleted successfully
...
  /scaling/policies:
    get:
      tags:
        - Scaling
      summary: List scaling policies
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All scaling policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScalingPolicy'
                  count:
                    type: integer
    post:
      tags:
        - Scaling
      summary: Create or replace scaling policy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScalingPolicy'
      responses:
        '201':
          description: Policy stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  policy:
                    $ref: '#/components/schemas/ScalingPolicy'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scaling/policies/{serviceId}:
    get:
      tags:
        - Scaling
      summary: Get scaling policy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      responses:
        '200':
          description: Scaling policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  policy:
                    $ref: '#/components/schemas/ScalingPolicy'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Scaling
      summary: Update scaling policy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScalingPolicy'
      responses:
        '200':
          description: Policy updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  policy:
                    $ref: '#/components/schemas/ScalingPolicy'
    delete:
      tags:
        - Scaling
      summary: Delete scaling policy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      responses:
        '200':
          description: Policy deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /scaling/services:
    get:
      tags:
        - Scaling
      summary: List service scaling states
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Per-service scaling state (replicas, last scale, cooldowns)
          content:
            application/json:
              schema:
                type: object
                properties:
                  services:
                    type: array
                    items:
                      type: object
  /scaling/services/{serviceId}:
    get:
      tags:
        - Scaling
      summary: Get service scaling state
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      responses:
        '200':
          description: Scaling state for one service
          content:
            application/json:
              schema:
                type: object
  /scaling/services/{serviceId}/history:
    get:
      tags:
        - Scaling
      summary: Scaling history
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      responses:
        '200':
          description: Scaling events for the service
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScalingEvent'
  /scaling/services/{serviceId}/scale:
    post:
      tags:
        - Scaling
      summary: Manual scale
      description: Immediately scale a service to a replica count.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ServiceIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - replicas
              properties:
                replicas:
                  type: integer
                  minimum: 1
                  maximum: 20
                reason:
                  type: string
      responses:
        '200':
          description: Scale event recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScalingEvent'
        '400':
          description: Invalid replica count or cooldown active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /scaling/status:
    get:
      tags:
        - Scaling
      summary: Autoscaler status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Whether the autoscaler is enabled plus counts
          content:
            application/json:
              schema:
                type: object
  /scaling/enable:
    post:
      tags:
        - Scaling
      summary: Enable autoscaler
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Autoscaler enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /scaling/disable:
    post:
      tags:
        - Scaling
      summary: Disable autoscaler
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Autoscaler disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /scaling/events:
    get:
      tags:
        - Scaling
      summary: Recent scaling events
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All recent scale up/down events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScalingEvent'

  # High availability
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Service ID
        project_id:
          type: string
          format: uuid
          description: Project ID
        name:
          type: string
          description: Service name
        type:
          type: string
          enum: [web, worker, database, cron]
          description: Service type
        status:
          type: string
          enum: [running, stopped, building, failed]
          description: Service status
        image:
          type: string
          description: Container image
        command:
          type: string
          description: Start command
        environment:
          type: string
          enum: [production, preview, development]
          description: Deployment environment
        git_repo:
          type: string
          description: Connected git repository
        git_branch:
          type: string
          description: Connected git branch
        build_path:
          type: string
          description: Build context path
        cpu:
          type: string
          description: CPU allocation
        memory:
          type: string
          description: Memory allocation
        replicas:
          type: integer
          description: Desired replica count
        port:
          type: integer
          description: Container port to expose (published host port, or Traefik when domain set)
        domain:
          type: string
          description: Public hostname routed via Traefik
        healthcheck_path:
          type: string
          description: HTTP path probed on port for container health
        restart_policy:
          type: string
          description: Docker restart policy (default unless-stopped)
        public_url:
          type: string
          description: Computed public URL (published port or domain)
        created_at:
          type: string
          format: date-time
          description: Service creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
...
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
          description: Service name
        type:
          type: string
          enum: [web, worker, database, cron]
          description: Service type
        image:
          type: string
          description: Container image
        command:
          type: string
          description: Start command
        environment:
          type: string
          enum: [production, preview, development]
          description: Deployment environment
        git_repo:
          type: string
          description: Git repository URL
        git_branch:
          type: string
          description: Git branch
        build_path:
          type: string
          description: Build context path
        cpu:
          type: string
          description: CPU allocation
        memory:
          type: string
          description: Memory allocation
        replicas:
          type: integer
          description: Desired replica count
        port:
          type: integer
          description: Container port to expose
        domain:
          type: string
          description: Public hostname routed via Traefik
        healthcheck_path:
          type: string
          description: HTTP path probed for container health
        restart_policy:
          type: string
          description: Docker restart policy
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    ScalingPolicy:
      type: object
      required:
        - service_id
        - min_replicas
        - max_replicas
      properties:
        service_id:
          type: string
        min_replicas:
          type: integer
          minimum: 1
        max_replicas:
          type: integer
          minimum: 1
        target_cpu:
          type: number
          description: Target CPU utilization percent
        target_memory:
          type: number
          description: Target memory utilization percent
        scale_up_cooldown:
          type: integer
          description: Nanoseconds (Go duration serialized as integer)
        scale_down_cooldown:
          type: integer
        scale_up_step:
          type: integer
        scale_down_step:
          type: integer
        metrics:
          type: array
          items:
            type: string
        thresholds:
          type: object
          additionalProperties:
            type: number
        enabled:
          type: boolean
        cost_optimization:
          type: object
    ScalingEvent:
      type: object
      properties:
        id:
          type: string
        service_id:
          type: string
        from_replicas:
          type: integer
        to_replicas:
          type: integer
        reason:
          type: string
        trigger:
          type: string
          description: auto | manual
        created_at:
          type: string
          format: date-time
...
  parameters:
...
    ServiceIdParam:
      name: serviceId
      in: path
      required: true
      description: Service ID
      schema:
        type: string
...